requirements.txt
selenium>=4.20.0
pandas>=2.2.0
numpy>=1.26.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
